ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
Pillow==10.2.0
numpy==1.26.3
torch==2.1.2
torchvision==0.16.2
transformers==4.37.0
scikit-learn==1.4.0
scipy==1.12.0
camel-tools==1.5.2
pyarabic==0.6.15
requests==2.31.0
httpx==0.26.0
redis==5.0.1
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
loguru==0.7.2
prometheus-client==0.19.0
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
aiofiles==23.2.1
